Migrate helpers into /lib
