+++ app/components/polaris/dropzone/component.rb
@@ -3,7 +3,7 @@
 module Polaris
   module Dropzone
     class Component < Polaris::Component
-      include Polaris::ActionHelper
+      include Polaris::Helpers::ActionHelper
 
       ALLOWED_TYPES = %w[file image]
