+++ app/components/polaris/text_field/component.rb
@@ -2,7 +2,7 @@
 module Polaris
   module TextField
     class Component < Polaris::Component
-      include Polaris::ActionHelper
+      include Polaris::Helpers::ActionHelper
 
       with_content_areas :connected_left, :connected_right
