--- app/helpers/polaris/conditional_helper.rb
@@ -1,11 +1,0 @@
-module Polaris
-  module ConditionalHelper
-    def conditional_wrapping_element(tag, options = {}, &block)
-      if options.delete(:show)
-        content_tag(tag, capture(&block), options)
-      else
-        capture(&block)
-      end
-    end
-  end
-end
