+++ lib/polaris/complex_action.rb
