+++ lib/polaris/helpers/conditional_helper.rb
@@ -1,0 +1,13 @@
+module Polaris
+  module Helpers
+    module ConditionalHelper
+      def conditional_wrapping_element(tag, options = {}, &block)
+        if options.delete(:show)
+          content_tag(tag, capture(&block), options)
+        else
+          capture(&block)
+        end
+      end
+    end
+  end
+end
